requirements.txt
cryptography==35.0.0
pycryptodome==3.11.0
PyNaCl==1.4.0
